Added support for PAPI events.
This took a little longer to complete since I had to test many things.
First, I tried hard making it work on Windows via Cygwin, but I did not manage
to make it work. Then I tried making it work on a Virtual Machine and even on
Windows Subsystem for Linux, but none of them worked either. The VM did not
have access to PAPI counters and WSL was limited to a very small amount of
events. In the end, I had to set-up a Linux machine.

Signed-off-by: Miguel Ángel Abella González <miguel.abella@miguelsoft.net>