Force usage of appropriate data types and values when creating arrays in benchmarks.
If no initialization value is specified, the lists will be initialized with zeroes of type integer. This will represent a problem when working with the lists as for some time there will be mixed data types (ints and floats), causing a possibly specialized list to become generic and thrashing possible performance optimizations.

Signed-off-by: Miguel Ángel Abella González <miguel.abella@miguelsoft.net>