Added support for proper PAPI multi-event handling.
The current implementation should work identically as in PolyBench/C. Please note that when running kernels in a loop the data arrays are not reinitialized again (same happens in C).

The benchmarks need to be rewritten for this change to take effect, but performance-wise no differences are expected meaning previously gathered results are still valid.

Signed-off-by: Miguel Ángel Abella González <miguel.abella@miguelsoft.net>